Decided to revert to enet 1.2.5 instead of enet 1.3.3.
(Rationale: enet 1.3 is not protocol compatible with 1.2, so users of
old versions of knights would not be able to connect to the new
server, if we used 1.3. They would not even get a message telling them
to upgrade; they would just get "connection failed" or some such. I
would prefer to avoid this situation. We can always upgrade to 1.3 at
a later date.)



git-svn-id: svn://www.knightsgame.org.uk/trunk@4 00979ad9-a3a6-4f7a-b49c-25ba5319b333