Fix item update bug when an observer connects mid-way through a game (Trac #115).
The problem was that because of the hack to use a temporary
DungeonView, the SEEN/UNSEEN cache inside the "real" ServerDungeonView
was not being updated, and so all squares were initially thought to be
UNSEEN. Therefore, the server would not send an update when the item
on a square was set to 0, because it thought the client had never seen
that square, and therefore (the server thought) the client already had
item 0 on that square.

Fixed by removing that hack. The "real" dungeonview is used for the
catchup, and therefore its SEEN/UNSEEN cache gets set correctly as
part of the initial catchup.  This does mean we send some unnecessary
updates to the existing players (when a new observer connects) but
there is no easy way to fix that at the moment.



git-svn-id: svn://www.knightsgame.org.uk/trunk@17 00979ad9-a3a6-4f7a-b49c-25ba5319b333