#185: On Linux (or on Windows when running the server), knights_data is now searched for in the current directory, or in DATA_DIR if not found in the current directory. (This can be overridden by using "-d" option to the client or editing the config file for the server.)
(On Windows when running the client, the game looks for
"knights_data" in the same dir as the Knights.exe, as before.)

We also now print the knights_data dir being used on startup.
This may help avoid confusion e.g. if people have a knights_data 
in both the current dir and /usr/local/share/knights.



git-svn-id: svn://www.knightsgame.org.uk/trunk@403 00979ad9-a3a6-4f7a-b49c-25ba5319b333